+++ src/lambdacd_git/ssh_agent_support.clj
@@ -1,6 +1,5 @@
 (ns lambdacd-git.ssh-agent-support
-  (:require [clojure.tools.logging :as log]
-            [lambdacd-git.ssh :as ssh])
+  (:require [clojure.tools.logging :as log])
   (:import (org.eclipse.jgit.transport SshSessionFactory JschConfigSessionFactory)
            (com.jcraft.jsch.agentproxy.connector SSHAgentConnector)
            (org.eclipse.jgit.util FS)
@@ -21,12 +20,3 @@
       (log/info "No SSH-Agent connector available. SSH-Keys with passphrases will not be supported"))
     (catch Exception e
       (log/warn e "Problems with SSH Agent. Falling back to default behavior"))))
-
-
-; INTERNAL: WILL BE MOVED TO ssh-init
-(defn session-factory [customizer-fns]
-  (ssh/session-factory customizer-fns))
-
-; DEPRECATED: USE lambdacd-git.ssh-init/init-ssh! instead!
-(defn initialize-ssh-agent-support! []
-  (SshSessionFactory/setInstance (ssh/session-factory [ssh-agent-customizer])))